refactor volumetric folder
