+++ nanome/_internal/volumetric/__init__.py
@@ -1,8 +1,4 @@
-from .unit_cell import _UnitCell
-from .volume_data import _VolumeData
-from .volume_layer import _VolumeLayer
-from .volume_properties import _VolumeProperties
+from .models import _UnitCell, _VolumeData, _VolumeLayer, _VolumeProperties
 
 # folders
 from . import io
-from . import serializers
