--- nanome/_internal/volumetric/unit_cell.py
@@ -1,13 +1,0 @@
-
-class _UnitCell(object):
-    def __init__(self):
-        from nanome.util import Vector3
-        self._A = 0.0
-        self._B = 0.0
-        self._C = 0.0
-
-        self._Alpha = 0.0
-        self._Beta = 0.0
-        self._Gamma = 0.0
-
-        self._Origin = Vector3()
