--- nanome/_internal/volumetric/volume_layer.py
@@ -1,7 +1,0 @@
-from nanome.util import Color
-
-
-class _VolumeLayer():
-    def __init__(self):
-        self._color = Color()
-        self._rmsd = 0.0
